build: make seeds idempotent
